requirements.txt
# Core LLM and API dependencies
openai>=1.63.0
requests>=2.25.0
python-dotenv>=0.19.0

# Data processing
pandas>=1.3.0
numpy>=1.20.0

# Logging and monitoring
wandb>=0.15.0
weave>=0.50.0

# Time zone handling
pytz>=2021.1

# Development and testing
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Type hints
typing-extensions>=4.0.0

# JSON handling (usually included in standard library)
# json - built-in

# Date/time handling
python-dateutil>=2.8.0

# Optional: For better CLI experience
click>=8.0.0
rich>=10.0.0

# Optional: For async support (if needed later)
aiohttp>=3.8.0 